refactor(model-agreement): extract aggregation helpers to keep resolver under file-size limit
The resolver hit 731 lines (limit 700). Moves all the pure synchronous helpers
(types, position-cell construction, comparable-cell collection, pair-summary
math, trial-consistency aggregation, empty-shape builders) into a new file
cloud/apps/api/src/services/model-agreement/aggregation.ts.

The resolver becomes 419 lines and only contains argument parsing, scope
resolution, snapshot reads, and Pothos plumbing. Pure aggregation logic is
now reusable and unit-testable in isolation.

No behavior change. All 27 tests still pass.
